% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allRand.R
\name{allRand}
\alias{allRand}
\title{Data Randomization}
\usage{
allRand(dataR, batchTot, numQC, withinN, numMatch, chkRep)
}
\arguments{
\item{dataR}{Data for randomization.}

\item{batchTot}{c(batchTot1, batchTot2) sizes of plates, just use one plate per batch, batch size inclusive of QC samples.}

\item{numQC}{Number of QC samples per batch.}

\item{withinN}{Number of samples away that the QC samples must be from each other.}

\item{numMatch}{Number of QC samples from a single mother within a batch.}

\item{chkRep}{Check if there is a repeat of the groups within the batches.}
}
\value{
A dataset with serum order randomized.
}
\description{
Randomizes aliquots to batches.
}
