% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{countQC}
\alias{countQC}
\title{Number of QC in each batch.}
\usage{
countQC(dataS, QCN)
}
\arguments{
\item{dataS}{Test dataset.}

\item{QCN}{Number of QC samples per batch.}
}
\value{
The output includes any batches that does not contain the number of QC samples specified.
}
\description{
Count how many QC samples are in each of the batches, and if it doesn't match the number specified.
}
