% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRand.R
\name{formatRand}
\alias{formatRand}
\title{Formats data}
\usage{
formatRand(QCdata, serumIDR, serumPack)
}
\arguments{
\item{serumIDR}{Serum data with serumIDs.}

\item{serumPack}{Serum data with packing lists.}

\item{QCData}{QC data.}
}
\value{
A dataset that is formatted and ready for the randomization file.
}
\description{
The function \code{formatRand} formats the dataset for the randomization function.  This function inputs serum data for both the study subjects and QC.
}
