% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{orderCases}
\alias{orderCases}
\title{Number of single group in sequence.}
\usage{
orderCases(dataI, betW)
}
\arguments{
\item{dataI}{Dataset to be tested.}

\item{betW}{Number of cases or controls to check if they are next to each other.}
}
\value{
The output stores if there are any cases or controls together beyond a certain specified value.
}
\description{
Tests if a large number of cases or controls are next to each other.
}
