% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randTest.R
\name{randTest}
\alias{randTest}
\title{QC Identifiers}
\usage{
randTest(dataMom, dataChild, maxAliq, nEvent)
}
\arguments{
\item{dataMom}{The mother dataset.}

\item{dataChild}{Child dataset.}

\item{maxAliq}{Number of aliquots per mother aliquot.}

\item{nEvent}{Number of aliquots per each of event or lab.}
}
\value{
The output of the function is a dataset with the ID links.
}
\description{
We first assign IDs linking mother/child and events using the function \code{randTest}.
}
