% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchR.R
\name{switchR}
\alias{switchR}
\title{Switching Generating Function}
\usage{
switchR(dataIn, numqc, numqcM, batchS)
}
\arguments{
\item{dataIn}{Randomized dataset.}

\item{numqc}{Number of QC samples per set.}

\item{numqcM}{Numberof QC matching samples.}

\item{batchS}{New batch size.}
}
\value{
A dataset with switches indicated.
}
\description{
Minimizes switches without completely re-randomizing the locations.
}
\examples{
serumSwitch <- switchR(dataIn=serumRand,numqc=2,numqcM=2,batchS=43)
}
