% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{testCCAcross}
\alias{testCCAcross}
\title{Sample groups within batches}
\usage{
testCCAcross(dataS)
}
\arguments{
\item{dataS}{The test dataset.}
}
\value{
The output lists all batches with not enough QC sample sets or the QC samples do not come from the same mother.
}
\description{
Tests and finds sample groups that are across batches.
}
