% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRand.R
\name{testRand}
\alias{testRand}
\title{Test Dataset}
\usage{
testRand(rowSize, colSize, studySize, expNS, numCC, QCpct, child)
}
\arguments{
\item{rowSize}{Max row size.}

\item{colSize}{Max column size.}

\item{studySize}{Number of cases.}

\item{expNS}{Number of aliquots per case/control.}

\item{numCC}{Number of controls per case.}

\item{QCpct}{Percent of QCs for number of samples.}

\item{child}{Number of children per "mother' aliquot.}
}
\value{
A practice dataset.
}
\description{
Generates test data for the randomization functions.
}
\examples{
testR <- testRand(rowSize=20,colSize=15,studySize=1000,expNS=7000,numCC=2,QCpct=0.05,child=4)
}
