% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_fcns.R
\name{uniqueID}
\alias{uniqueID}
\title{Unique IDs}
\usage{
uniqueID(testD, IDN)
}
\arguments{
\item{testD}{Test dataset.}

\item{IDN}{ID to test.}
}
\value{
Any IDs that are not unique.
}
\description{
Test for unique IDs.
}
